+++ client/src/common/actions.ts
@@ -1,0 +1,8 @@
+import * as vscode from 'vscode';
+import { LanguageClient } from 'vscode-languageclient/node';
+
+
+
+export function registerActions(context: vscode.ExtensionContext, client: LanguageClient): void {
+	
+}