+++ client/src/common/constants.ts
@@ -19,11 +19,14 @@ export const COMMANDS = {
 	USER_API_KEY: `${EXTENSION_ID}.apiKey`,
 	REMOVE_API_KEY: `${EXTENSION_ID}.removeApiKey`,
 	GET_API_KEY: `${EXTENSION_ID}.getApiKey`,
+	PROVIDE_EXCEPTION_HANDLING: `${EXTENSION_ID}.provideExceptionHandling`,
+	ANALYZE_EXCEPTION_HANDLING: `${EXTENSION_ID}.analyzeExceptionHandling`,
+	PREVIEW_AND_APPLY_SUGGESTION: `${EXTENSION_ID}.previewAndApplySuggestion`,
 };
 
 export const RATE_LIMIT_NOTIFICATION_ID = `${EXTENSION_ID}/rateLimitReached`;
 export const ACCESS_FORBIDDEN_NOTIFICATION_ID = `${EXTENSION_ID}/forbidden`;
 
-export const API_SERVER_URL = process.env.NODE_ENV === "development" ?  "http://localhost:8000" : "https://rome-django-auth.onrender.com";
+export const API_SERVER_URL = process.env.NODE_ENV === "development" ?  "http://localhost:8000" : "https://djangoly-api.onrender.com";
 
 export const API_KEY_SIGNUP_URL = "https://forms.gle/gEEZdfhWpQyQh2qVA";