+++ client/src/common/constants/errors.ts
@@ -1,0 +1,7 @@
+export enum ERROR_CODES  {
+    UNAUTHENTICATED = 403,
+}
+
+export const ERROR_MESSAGES = {
+	FEATURE_GATED: 'You do not have access to this feature. Please sign in or sign up for an account.',
+};