+++ client/src/common/constants/messages.ts
@@ -1,0 +1,11 @@
+
+export enum AUTH_MESSAGES  {
+    FREE_API_KEY_PROMPT = "Djangoly is currently in free beta, but a valid API key is required to continue. Don't have an API key? Request one by completing the form below.",
+    // TODO: Add links to our privacy policy and terms of service
+    WELCOME_BETA_MESSAGE = "Welcome to Djangoly (Beta)! 👋 By using this extension, you agree to our Terms of Service and Privacy Policy.",
+    INVALID_API_KEY = "The API key you entered is invalid. Please try again or request a new API key using the form below.",
+    AUTHENTICATION_REQUIRED = "Authentication is required to use Djangoly. Please sign in or request an API key.",
+    SIGN_OUT = "You have deactivated Djangoly. To reactivate, reload the current window or restart your IDE. See ya! 👋",
+    GITHUB_SIGN_IN = "Sign in to continue using Djangoly. By using this extension you agree to our Terms of Service and Privacy Policy.",
+    MUST_AGREE_TO_TERMS = "You must accept the Terms of Service and Privacy Policy to use Djangoly."
+}