+++ server/src/constants/commands.ts
@@ -10,6 +10,7 @@ export const CREATE_REPOSITORY = `${EXTENSION_ID}.createRepository`;
 export const CHECK_TESTS_EXISTS = `${EXTENSION_ID}.checkTestExists`;
 export const UPDATE_CACHED_USER_TOKEN = `${EXTENSION_ID}.updateCachedUserToken`;
 export const REPORT_FALSE_POSITIVE = `${EXTENSION_ID}.reportFalsePositive`;
+export const PROVIDE_EXCEPTION_HANDLING = `${EXTENSION_ID}.provideExceptionHandling`;
 
 export const RATE_LIMIT_NOTIFICATION_ID = `${EXTENSION_ID}/rateLimitReached`;
 export const ACCESS_FORBIDDEN_NOTIFICATION_ID = `${EXTENSION_ID}/forbidden`;
@@ -20,7 +21,8 @@ const COMMANDS = {
 	CREATE_REPOSITORY,
 	CHECK_TESTS_EXISTS,
 	UPDATE_CACHED_USER_TOKEN,
-	REPORT_FALSE_POSITIVE
+	REPORT_FALSE_POSITIVE,
+	PROVIDE_EXCEPTION_HANDLING
 };
 
 export const COMMANDS_LIST = Object.values(COMMANDS);